+++ src/DodoCMS/ViewModels/Role.cs
@@ -27,11 +27,11 @@ namespace Dodo.ViewModels
 		public Resources.ResourceReference<Resources.IRESTResource> Parent { get; set; }
 		[DisplayName("Public Description")]
 		[View(EPermissionLevel.PUBLIC, EPermissionLevel.ADMIN)]
-		[Resources.MaxStringLengthAttribute(2048)]
+		[Resources.MaxStringLengthAttribute((int)2048)]
 		public string PublicDescription { get; set; }
 		[DisplayName("Applicant Question")]
 		[View(EPermissionLevel.PUBLIC, EPermissionLevel.ADMIN)]
-		[Resources.MaxStringLengthAttribute(2048)]
+		[Resources.MaxStringLengthAttribute((int)2048)]
 		public string ApplicantQuestion { get; set; }
 		[DisplayName("Published")]
 		[View(EPermissionLevel.ADMIN, EPermissionLevel.ADMIN)]
